feat: add handler for grant updation from grantFactory contract
