+++ src/grant-mapping.ts
@@ -1,14 +1,15 @@
 import { BigInt, log } from '@graphprotocol/graph-ts'
-import { GrantCreated } from '../generated/QBGrantFactoryContract/QBGrantFactoryContract'
+import { GrantCreated, GrantUpdatedFromFactory } from '../generated/QBGrantFactoryContract/QBGrantFactoryContract'
 import { ApplicationMilestone, FundsTransfer, Grant, GrantApplication, Workspace } from '../generated/schema'
 import { QBGrantsContract } from '../generated/templates'
 import { DisburseReward, DisburseRewardFailed, FundsDepositFailed, FundsWithdrawn, GrantUpdated, TransactionRecord } from '../generated/templates/QBGrantsContract/QBGrantsContract'
 import { validatedJsonFromIpfs } from './json-schema/json'
 import { applyGrantFundUpdate } from './utils/apply-grant-deposit'
-import { dateToUnixTimestamp, isPlausibleIPFSHash, mapGrantFieldMap, mapGrantManagers, mapGrantRewardAndListen, removeEntityCollection } from './utils/generics'
+import { dateToUnixTimestamp, mapGrantFieldMap, mapGrantManagers, mapGrantRewardAndListen } from './utils/generics'
 import { disburseReward } from './utils/handle-disburse-reward'
+import handleGrantUpdate from './utils/handle-grant-update'
 import { addFundsTransferNotification } from './utils/notifications'
-import { GrantCreateRequest, GrantUpdateRequest, validateGrantCreateRequest, validateGrantUpdateRequest } from './json-schema'
+import { GrantCreateRequest, validateGrantCreateRequest } from './json-schema'
 
 export function handleGrantCreated(event: GrantCreated): void {
 	const workspaceId = event.params.workspaceId.toHex()
@@ -121,56 +122,11 @@ export function handleFundsWithdrawn(event: FundsWithdrawn): void {
 
 export function handleGrantUpdated(event: GrantUpdated): void {
 	const grantId = event.transaction.to!.toHex()
+	handleGrantUpdate(grantId, event)
+}
 
-	const entity = Grant.load(grantId)
-	if(!entity) {
-		log.warning(`[${event.transaction.hash.toHex()}] recv grant update for unknown grant, ID="${grantId}"`, [])
-		return
-	}
+export function handleGrantUpdatedFromFactory(event: GrantUpdatedFromFactory) {
+	const grantId = event.params.grantAddress.toHex()
 
-	entity.updatedAtS = event.params.time.toI32()
-	entity.workspace = event.params.workspaceId.toHex()
-	entity.acceptingApplications = event.params.active
-
-	const hash = event.params.metadataHash
-	if(isPlausibleIPFSHash(hash)) {
-		const jsonResult = validatedJsonFromIpfs<GrantUpdateRequest>(hash, validateGrantUpdateRequest)
-		if(jsonResult.error) {
-			log.warning(`[${event.transaction.hash.toHex()}] error in updating grant metadata, error: ${jsonResult.error!}`, [])
-			return
-		}
-
-		const json = jsonResult.value!
-		if(json.title) {
-			entity.title = json.title!
-		}
-
-		if(json.summary) {
-			entity.summary = json.summary!
-		}
-
-		if(json.details) {
-			entity.details = json.details!
-		}
-
-		if(json.deadline) {
-			entity.deadline = json.deadline!.toISOString()
-			entity.deadlineS = dateToUnixTimestamp(json.deadline!)
-		}
-
-		if(json.reward) {
-			entity.reward = mapGrantRewardAndListen(entity.id, entity.workspace, json.reward!).id
-		}
-
-		if(json.fields) {
-			entity.fields = mapGrantFieldMap(entity.id, json.fields!)
-		}
-
-		if(json.grantManagers && json.grantManagers!.length) {
-			removeEntityCollection('GrantManager', entity.managers)
-			entity.managers = mapGrantManagers(json.grantManagers, entity.id, entity.workspace)
-		}
-	}
-
-	entity.save()
+	handleGrantUpdate(grantId, event)
 }
